Updated documentation and added code for eusart1's (bluetooth to the remote) receive interrupt. Added skeletons for audio controller and update IR header to include a struct to represent packets. Extensively modified the LED logic functions in hopes that the optimized compiler will be able to generate code that will allow for more flexible control of the LEDs without having timing that is out of spec.

+++ Code/Mine/MagicMineTesting.X/util_functions.c
@@ -8,7 +8,8 @@
 
 #include "util_functions.h"
 
-void delay_25ms_n_times(uint8_t n) {
+void delay_25ms_n_times(uint16_t n) {
+    
     for (n; n > 1; n--) {
         __delay_ms (25);
     }
